+++ Source/SelfService/Web/applications/containerregistry/index.tsx
@@ -13,7 +13,7 @@ import { WorkSpaceLayoutWithSidePanel } from '../../layout/workSpaceLayout';
 import { RegistryContainer } from './container/registryContainer';
 import { RouteNotFound } from '../../components/notfound';
 
-import { withRouteApplicationState } from '../../spaces/applications/withRouteApplicationState';
+import { withRouteApplicationState } from '../../utils/withRouteApplicationState';
 
 export const ContainerRegistryIndex = withRouteApplicationState(({ routeApplicationParams }) => {
     const navigate = useNavigate();
