+++ Source/SelfService/Web/components/applicationBuilding.tsx
